+++ src/config.rs
@@ -11,8 +11,8 @@ pub struct CameraConfig {
     pub vup: Vec<f64>,
     pub samples_per_pixel: i32,
     pub max_ray_bounces: i32,
-    pub vertical_field_of_view: i32,
-    pub defocus_angle: i32,
+    pub vertical_field_of_view: f64,
+    pub defocus_angle: f64,
     pub focus_distance: f64,
 }
